feat: create InspectorClient lib to separate Entra and database inspector queries
