+++ apps/web/src/app/inspector/inspector.test.js
@@ -1,12 +1,6 @@
 import { describe, it, mock } from 'node:test';
 import { strict as assert } from 'node:assert';
-import {
-	getInspectorById,
-	fetchInspectorList,
-	getSortedInspectorList,
-	getInspectorList,
-	getInspectorDetails
-} from './inspector.js';
+import { getInspectorById, fetchInspectorList, getSortedInspectorList, getInspectorList } from './inspector.js';
 
 const groupId = 'groupId';
 const mockSession = {};
@@ -446,35 +440,5 @@ describe('inspectors', () => {
 		mockEntraClient.listAllGroupMembers.mock.mockImplementationOnce(() => groupMemberList);
 		const inspectorList = await getInspectorList(mockService, mockSessionWithAccount);
 		assert.deepStrictEqual(inspectorList, expectedInspectorList);
-	});
-});
-
-describe('getInspectorDetails', () => {
-	it('should return null if no inspectorId is provided', async () => {
-		const mockDb = {
-			inspectors: {
-				findUnique: mock.fn()
-			}
-		};
-		const result = await getInspectorDetails(mockDb, undefined);
-		assert.equal(result, null);
-	});
-	it('should return inspector if inspectorId is provided', async () => {
-		const mockInspector = {
-			id: '1',
-			firstName: 'John',
-			lastName: 'Doe'
-		};
-		const mockDb = {
-			inspector: {
-				findFirst: mock.fn(() => mockInspector)
-			}
-		};
-		const result = await getInspectorDetails(mockDb, 'entra-id-1');
-		assert.equal(result, mockInspector);
-		assert.equal(mockDb.inspector.findFirst.mock.callCount(), 1);
-		const args = mockDb.inspector.findFirst.mock.calls[0].arguments[0];
-		assert.deepEqual(args.where?.entraId, 'entra-id-1');
-		assert.deepEqual(args?.include, { Specialisms: true });
 	});
 });
